app/requirements.txt
numpy==1.18.4
pandas==1.0.4
python-dateutil==2.8.1
pytube3==9.6.4
pytz==2020.1
six==1.15.0
srt==3.3.0
typing-extensions==3.7.4.2
